# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the ShelterTech SF Service Guide application.",
        "title": "ShelterTech API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api",
    "paths": {
...
        "/resources/count": {
            "get": {
                "description": "get the total count of resources",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get Resource Count",
                "responses": {
                    "200": {
                        "description": "Resource count",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Error"
                        }
                    }
                }
            }
        },
        "/resources/{id}": {
            "get": {
                "description": "gets a single resource by resource ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get Resource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resources.ResourceResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "addresses.Address": {
            "type": "object",
            "properties": {
                "address_1": {
                    "type": "string"
                },
                "address_2": {
                    "type": "string"
                },
                "address_3": {
                    "type": "string"
                },
                "address_4": {
                    "type": "string"
                },
                "attention": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "state_province": {
                    "type": "string"
                }
            }
        },
...
        "categories.Category": {
            "type": "object",
            "properties": {
                "featured": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "top_level": {
                    "type": "boolean"
                }
            }
        },
...
        "common.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "documents.Document": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "eligibilities.Eligibility": {
            "type": "object",
            "properties": {
                "feature_rank": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "instructions.Instruction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "instruction": {
                    "type": "string"
                }
            }
        },
...
        "notes.Note": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "phones.Phone": {
            "type": "object",
            "properties": {
                "country_code": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                },
                "service_type": {
                    "type": "string"
                }
            }
        },
...
        "resources.Resource": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/addresses.Address"
                    }
                },
                "alternate_name": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/categories.Category"
                    }
                },
                "certified": {
                    "type": "boolean"
                },
                "certified_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "featured": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "internal_note": {
                    "type": "string"
                },
                "legal_status": {
                    "type": "string"
                },
                "long_description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notes.Note"
                    }
                },
                "phones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/phones.Phone"
                    }
                },
                "schedule": {
                    "$ref": "#/definitions/schedules.Schedule"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resources.ResourceService"
                    }
                },
                "short_description": {
                    "type": "string"
                },
                "source_attribution": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "resources.ResourceResponse": {
            "type": "object",
            "properties": {
                "resource": {
                    "$ref": "#/definitions/resources.Resource"
                }
            }
        },
        "resources.ResourceService": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/addresses.Address"
                    }
                },
                "alternate_name": {
                    "type": "string"
                },
                "application_process": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/categories.Category"
                    }
                },
                "certified": {
                    "type": "boolean"
                },
                "certified_at": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/documents.Document"
                    }
                },
                "eligibilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/eligibilities.Eligibility"
                    }
                },
                "eligibility": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "featured": {
                    "type": "boolean"
                },
                "fee": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/instructions.Instruction"
                    }
                },
                "internal_note": {
                    "type": "string"
                },
                "interpretation_services": {
                    "type": "string"
                },
                "long_description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notes.Note"
                    }
                },
                "required_documents": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/schedules.Schedule"
                },
                "short_description": {
                    "type": "string"
                },
                "source_attribution": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                },
                "wait_time": {
                    "type": "string"
                }
            }
        },
...
        "schedules.Schedule": {
            "type": "object",
            "properties": {
                "hours_known": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "schedule_days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schedules.ScheduleDay"
                    }
                }
            }
        },
        "schedules.ScheduleDay": {
            "type": "object",
            "properties": {
                "close_day": {
                    "type": "string"
                },
                "close_time": {
                    "type": "string"
                },
                "closes_at": {
                    "type": "integer"
                },
                "day": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "open_day": {
                    "type": "string"
                },
                "open_time": {
                    "type": "string"
                },
                "opens_at": {
                    "type": "integer"
                }
            }
        },
...
    }
}